+++ src/main/java/net/brnbrd/delightful/common/item/food/CupItem.java
@@ -7,8 +7,11 @@ import net.minecraft.world.item.Items;
 public class CupItem extends CompatConsumableItem {
 
 	public CupItem(Properties properties, boolean hasFoodEffectTooltip, boolean hasCustomTooltip) {
-		super(properties.stacksTo(16)
-				.craftRemainder(Util.item(Util.rl(Mods.MD, "copper_cup"), Items.BOWL)),
-			hasFoodEffectTooltip, hasCustomTooltip, Mods.MD);
+		super(
+				properties.stacksTo(16).craftRemainder(Util.item(Util.rl(Mods.MD, "copper_cup"), Items.BOWL)),
+				hasFoodEffectTooltip,
+				hasCustomTooltip,
+				Mods.MD
+		);
 	}
 }