+++ src/main/java/net/brnbrd/delightful/data/gen/DelightfulRecipeProvider.java
@@ -390,7 +390,7 @@ public class DelightfulRecipeProvider extends RecipeProvider implements IConditi
 						.unlockedBy("has_nuts", has(DelightfulItemTags.NUTS)),
 				"food/cooking/nut_butter_bottle", finished, enabled(DelightfulItems.NUT_BUTTER_BOTTLE), not(tagEmpty(DelightfulItemTags.NUTS)), not(modLoaded(Mods.VD)));
 		wrap(CookingPotRecipeBuilder.cookingPotRecipe(
-								DelightfulItems.CACTUS_CHILI.get(), 1, CookingRecipes.NORMAL_COOKING, 0.35F, Items.BOWL)
+								DelightfulItems.CACTUS_CHILI.get(), 1, CookingRecipes.NORMAL_COOKING, 1.0F, Items.BOWL)
 						.addIngredient(ForgeTags.RAW_PORK)
 						.addIngredient(ModItems.TOMATO_SAUCE.get())
 						.addIngredient(DelightfulItems.CACTUS_FLESH.get())
@@ -399,7 +399,7 @@ public class DelightfulRecipeProvider extends RecipeProvider implements IConditi
 						.setRecipeBookTab(CookingPotRecipeBookTab.MEALS),
 				"food/cooking/cactus_chili", finished, enabled(DelightfulItems.CACTUS_CHILI), enabled(DelightfulItems.CACTUS_FLESH));
 		wrap(CookingPotRecipeBuilder.cookingPotRecipe(
-								DelightfulItems.CACTUS_SOUP.get(), 1, CookingRecipes.NORMAL_COOKING, 0.35F, Items.BOWL)
+								DelightfulItems.CACTUS_SOUP.get(), 1, CookingRecipes.NORMAL_COOKING, 1.0F, Items.BOWL)
 						.addIngredient(DelightfulItems.CACTUS_FLESH.get())
 						.addIngredient(ForgeTags.VEGETABLES_TOMATO)
 						.addIngredient(DelightfulItemTags.HOT_SPICE)
@@ -408,7 +408,16 @@ public class DelightfulRecipeProvider extends RecipeProvider implements IConditi
 						.setRecipeBookTab(CookingPotRecipeBookTab.MEALS),
 				"food/cooking/cactus_soup", finished, enabled(DelightfulItems.CACTUS_SOUP), enabled(DelightfulItems.CACTUS_FLESH));
 		wrap(CookingPotRecipeBuilder.cookingPotRecipe(
-								DelightfulItems.COCONUT_CURRY.get(), 1, CookingRecipes.NORMAL_COOKING, 0.35F)
+								DelightfulItems.VENISON_STEW.get(), 1, CookingRecipes.NORMAL_COOKING, 1.0F, Items.BOWL)
+						.addIngredient(DelightfulItemTags.RAW_VENISON)
+						.addIngredient(ForgeTags.VEGETABLES_CARROT)
+						.addIngredient(ForgeTags.VEGETABLES_POTATO)
+						.addIngredient(ForgeTags.VEGETABLES_POTATO)
+						.unlockedBy("has_raw_venison", has(DelightfulItemTags.RAW_VENISON))
+						.setRecipeBookTab(CookingPotRecipeBookTab.MEALS),
+				"food/cooking/venison_stew", finished, enabled(DelightfulItems.VENISON_STEW), not(tagEmpty(DelightfulItemTags.RAW_VENISON)));
+		wrap(CookingPotRecipeBuilder.cookingPotRecipe(
+								DelightfulItems.COCONUT_CURRY.get(), 1, CookingRecipes.NORMAL_COOKING, 1.0F, Items.BOWL)
 						.addIngredient(DelightfulItemTags.COCONUT)
 						.addIngredient(ForgeTags.MILK)
 						.addIngredient(ForgeTags.RAW_CHICKEN)
@@ -419,7 +428,7 @@ public class DelightfulRecipeProvider extends RecipeProvider implements IConditi
 						.setRecipeBookTab(CookingPotRecipeBookTab.MEALS),
 				"food/cooking/coconut_curry_no_ginger", finished, enabled(DelightfulItems.COCONUT_CURRY), not(tagEmpty(DelightfulItemTags.COCONUT)), tagEmpty(DelightfulItemTags.VEGETABLES_GINGER));
 		wrap(CookingPotRecipeBuilder.cookingPotRecipe(
-								DelightfulItems.COCONUT_CURRY.get(), 1, CookingRecipes.NORMAL_COOKING, 0.35F)
+								DelightfulItems.COCONUT_CURRY.get(), 1, CookingRecipes.NORMAL_COOKING, 1.0F, Items.BOWL)
 						.addIngredient(DelightfulItemTags.COCONUT)
 						.addIngredient(ForgeTags.MILK)
 						.addIngredient(ForgeTags.RAW_CHICKEN)
@@ -430,7 +439,7 @@ public class DelightfulRecipeProvider extends RecipeProvider implements IConditi
 						.setRecipeBookTab(CookingPotRecipeBookTab.MEALS),
 				"food/cooking/coconut_curry", finished, enabled(DelightfulItems.COCONUT_CURRY), not(tagEmpty(DelightfulItemTags.COCONUT)), not(tagEmpty(DelightfulItemTags.VEGETABLES_GINGER)));
 		wrap(CookingPotRecipeBuilder.cookingPotRecipe(
-								DelightfulItems.SINIGANG.get(), 1, CookingRecipes.NORMAL_COOKING, 0.35F)
+								DelightfulItems.SINIGANG.get(), 1, CookingRecipes.NORMAL_COOKING, 1.0F, Items.BOWL)
 						.addIngredient(ForgeTags.RAW_FISHES)
 						.addIngredient(ModItems.TOMATO_SAUCE.get())
 						.addIngredient(ForgeTags.VEGETABLES_ONION)
@@ -441,7 +450,7 @@ public class DelightfulRecipeProvider extends RecipeProvider implements IConditi
 						.setRecipeBookTab(CookingPotRecipeBookTab.MEALS),
 				"food/cooking/sinigang", finished, enabled("sinigang"), not(tagEmpty(DelightfulItemTags.VEGETABLES_GINGER)), not(tagEmpty(DelightfulItemTags.FRUITS_CITRUS)));
 		wrap(CookingPotRecipeBuilder.cookingPotRecipe(
-								DelightfulItems.SINIGANG.get(), 1, CookingRecipes.NORMAL_COOKING, 0.35F)
+								DelightfulItems.SINIGANG.get(), 1, CookingRecipes.NORMAL_COOKING, 1.0F, Items.BOWL)
 						.addIngredient(ForgeTags.RAW_PORK)
 						.addIngredient(ForgeTags.RAW_FISHES)
 						.addIngredient(ForgeTags.GRAIN_RICE)